requirements.txt
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Django==1.11.7
djangorestframework==3.7.3
easygui==0.98.1
httpie==0.9.9
idna==2.6
matplotlib==2.1.0
numpy==1.13.3
opencv-python==3.3.0.10
pandas==0.21.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
